价格是小数的问题

+++ project/wisdom-project/src/main/resources/static/customer/js/native/orderPay.js
@@ -196,7 +196,7 @@ var orderPayInit = function(){
                         $('#needPayOrderList').html(needPayOrderListHtml);
 
                         $.each(needPayOrderList,function(index,value){
-                            totalPrice = totalPrice + parseInt(value.productNum)*parseInt(value.productPrice);
+                            totalPrice = totalPrice + parseFloat(value.productNum)*parseFloat(value.productPrice);
                             trainingProductId = value.productId;
                             orderIds.push(needPayOrderList[index].orderId);
                         })
